requirements.txt
boto==2.43.0
boto3==1.4.1
Flask==0.11
flask_mongoengine==0.7.5
flask_login==0.3.2
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
chardet==2.1.1
requests==2.2.1
pylint
coverage
Fabric3
